mina-tx-type: add N type, ToChecked, SignedRandExt, SlotRandExt
Moves additional types and traits from ledger to mina-tx-type:

- N type: Generic 64-bit number for proof computations
- ToChecked trait: Convert currency types to checked circuit forms
- SignedRandExt trait: Random signed value generation
- SlotRandExt trait: Random slot generation for tests

Also consolidates ToInputs implementations in ledger/src/hash.rs using
the impl_to_inputs macro for currency types.

Note: ToInputs trait cannot be moved to mina-tx-type due to Rust's
orphan rule - ledger needs to implement ToInputs for foreign types.

Issue: #1665
